requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
distlib==0.3.9
Django==5.1.2
django-appconf==1.0.6
django-cors-headers==4.5.0
django-filter==24.3
django-rest-knox==5.0.2
djangorestframework==3.15.2
drf-yasg==1.21.7
Faker==30.3.0
filelock==3.16.1
inflection==0.5.1
Markdown==3.7
packaging==24.1
platformdirs==4.3.6
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
virtualenv==20.26.6
